Fix: pdf 업로드 예외처리
1. DB상 파일명이 그대로이나 물리적 파일을 삭제했을 경우
2. 파일을 여러 번 등록할 경우
를 대비함.

+++ src/main/java/com/dominest/dominestbackend/domain/resident/ResidentService.java
@@ -33,24 +33,23 @@ public class ResidentService {
     /** @return 저장한 파일명 */
     @Transactional
     public String uploadPdf(Long id, FileService.FilePrefix filePrefix, MultipartFile pdf) {
+        Resident resident = findById(id);
         // 로컬에 파일 저장
         String uploadedFileName = fileService.save(filePrefix, pdf);
-        Resident resident = findById(id);
         String prevFileName = resident.getPdfFileName();
 
         // 파일명 저장 후 반환
         resident.setPdfFileName(uploadedFileName);
 
-        fileService.deleteFile(FileService.FilePrefix.RESIDENT_PDF, prevFileName);
+        if (prevFileName != null)
+            fileService.deleteFile(FileService.FilePrefix.RESIDENT_PDF, prevFileName);
+
         return uploadedFileName;
     }
 
-    // Todo 무시된 데이터는 클라이언트에 반환하기. 'xxx.pdf' 파일이 변경되었다 식으로.
-    /**@return 업로드한 파일 개수*/
     @Transactional
     public PdfBulkUploadDto.Res uploadPdfs(FileService.FilePrefix filePrefix, List<MultipartFile> pdfs, ResidenceSemester residenceSemester) {
         PdfBulkUploadDto.Res res = new PdfBulkUploadDto.Res();
-
         for (MultipartFile pdf : pdfs) {
             // 빈 객체면 continue
             if (pdf.isEmpty()) {
@@ -76,10 +75,15 @@ public class ResidentService {
             // 로컬에 파일 저장
             String uploadedFileName = fileService.save(filePrefix, pdf);
 
-            // resident 객체에 파일명 저장
+            String prevFileName = resident.getPdfFileName();
+            // 파일명 저장 후 반환
             resident.setPdfFileName(uploadedFileName);
+
             res.addToDtoList(filename, "OK", null);
             res.addSuccessCount();
+
+            if (prevFileName != null)
+                fileService.deleteFile(FileService.FilePrefix.RESIDENT_PDF, prevFileName);
         }
         // 한 건도 업로드하지 못했으면 예외발생
         if (res.getSuccessCount() == 0)
